Add files via upload
In Data Wrangling I learned how to use various types of data to make analysis and display it in a way that makes discoveries easy for everyone to see. Most importantly, I learned how to find data and manipulate it to be usable. The first part of my project involved analyzing a dataset on NBA players. After examining the data, I wanted to see where in the USA the most talented NBA players come from. The second part used data from Wikipedia about the highest-grossing movies online, and I cleaned the data to display movies based on how important the budget was to how high the movies grossed.