Add files via upload
In Bayesian Statistics, I learned how to use the powerful tool of Bayes rule to create interesting conclusions from data. This branch of statistics is incredibly valuable because in Bayesian statistics, probabilities are used to quantify uncertainty. Unlike classical statistics, which uses fixed probabilities to model events, Bayesian statistics uses probabilities that are updated based on new information. This means that prior knowledge, beliefs, or assumptions about a parameter or a model can be incorporated into the analysis. In my final project, I applied Bayesian analysis to my own data.