Add files via upload
Regression methods are a statistical technique used to model the relationship between a dependent variable and one or more independent variables. It is a powerful tool for predicting outcomes and understanding the factors that influence them. After learning about Regression methods, my colleagues and I created a thorough model building investigation on a Chicago Redline dataset we found to make predictions on the FAIR plan renewal. This project allowed us to gain hands-on experience in applying regression methods to real-world data and understanding how to interpret the results.